+++ src/Model/YamlServiceAliasing/YamlServiceAliasingChecker.php
@@ -6,7 +6,7 @@ namespace YamlStandards\Model\YamlServiceAliasing;
 
 use SebastianBergmann\Diff\Differ;
 use YamlStandards\Command\ProcessOutput;
-use YamlStandards\Model\CheckerInterface;
+use YamlStandards\Model\AbstractChecker;
 use YamlStandards\Model\Component\YamlService;
 use YamlStandards\Model\Config\StandardParametersData;
 use YamlStandards\Result\Result;
@@ -14,7 +14,7 @@ use YamlStandards\Result\Result;
 /**
  * Check service file observe uniform aliasing
  */
-class YamlServiceAliasingChecker implements CheckerInterface
+class YamlServiceAliasingChecker extends AbstractChecker
 {
     /**
      * @inheritDoc