+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Yaml files path service: use wildcards for find files instead of directory recursive iterator
+- Yaml empty line at end: this standard can be used for any file suffix
+
 ### Fixed
 - Yaml indent: fix get correct indents for array
 ```diff